refactor(TransferManagerWorker): Moved to a dedicated file
